Pack readTagColors + modificationColors as ABGR u32 in RPC worker
The RPC worker now emits one u32 per read/modification instead of
3-or-4 loose bytes. The main thread writes a single UBO/vertex slot
instead of packing bytes each render frame, and the read instance
stride drops 68B → 60B.

- buildTagColors / extractModifications in processFeatureAlignments
  produce Uint32Array, packing via (r | g<<8 | b<<16 | a<<24).
- read.slang now declares `uint tagColor : ATTR9` and unpacks at the
  one call site via `unpackRGBA(inst.tagColor).xyz`. The "tag color
  set?" check collapses from a 3-channel non-zero test to `!= 0u`.
- PileupDataResult, ReadUploadData, ModificationUploadData, and the
  Canvas2D mirror types all narrow to Uint32Array.
- Canvas2D + SVG renderers use the shared abgrRed/Green/Blue/Alpha /
  abgrToCssRgba helpers from packages/core/src/util/colorBits.ts.
- Test fixtures updated to Uint32Array(0).

+++ plugins/alignments/src/LinearAlignmentsDisplay/components/rendererTypes.ts
@@ -48,7 +48,7 @@ export interface ReadUploadData {
   readInsertSizes: Float32Array
   readPairOrientations: Uint8Array
   readStrands: Int8Array
-  readTagColors: Uint8Array
+  readTagColors: Uint32Array
   readChainHasSupp?: Uint8Array
   readIds: string[]
   numReads: number
@@ -86,7 +86,7 @@ export interface CigarUploadData {
 export interface ModificationUploadData {
   modificationPositions: Uint32Array
   modificationYs: Uint16Array
-  modificationColors: Uint8Array
+  modificationColors: Uint32Array
   numModifications: number
 }
