+++ plugins/alignments/src/shared/extractFeatureArrays.ts
@@ -121,7 +121,7 @@ export function extractFeatureArrays<T extends FeatureData>(
 
   const tagColors = isTagColorMode
     ? buildTagColors(features, tagColorValues, colorBy, colorTagMap)
-    : new Uint8Array(0)
+    : new Uint32Array(0)
 
   modifications.sort((a, b) => a.modType.localeCompare(b.modType))
